+++ tests/port4me.R
@@ -1,11 +1,85 @@
 library(port4me)
 
-Sys.setenv(PORT4ME_USER = "alice")
+message('- port4me(user = "alice")')
+truth <- 30845L
+ports <- port4me(user = "alice")
+print(ports)
+stopifnot(
+  is.integer(ports),
+  all(is.finite(ports)),
+  all(ports > 0L),
+  all(ports <= 65535L),
+  all(ports >= 1024L),
+  length(ports) == length(truth),
+  all(ports == truth)
+)
+
+
+message('- port4me(user = "bob")')
+truth <- 54242L
+ports <- port4me(user = "bob")
+print(ports)
+stopifnot(
+  is.integer(ports),
+  all(is.finite(ports)),
+  all(ports > 0L),
+  all(ports <= 65535L),
+  all(ports >= 1024L),
+  length(ports) == length(truth),
+  all(ports == truth)
+)
 
 
 message('- port4me(user = "alice", tool = "rstudio")')
 truth <- 22486L
 ports <- port4me(user = "alice", tool="rstudio")
+print(ports)
+stopifnot(
+  is.integer(ports),
+  all(is.finite(ports)),
+  all(ports > 0L),
+  all(ports <= 65535L),
+  all(ports >= 1024L),
+  length(ports) == length(truth),
+  all(ports == truth)
+)
+
+
+message('- port4me(user = "alice") with PORT4ME_TOOL=rstudio)')
+Sys.setenv(PORT4ME_TOOL = "rstudio")
+truth <- 22486L
+ports <- port4me(user = "alice")
+print(ports)
+stopifnot(
+  is.integer(ports),
+  all(is.finite(ports)),
+  all(ports > 0L),
+  all(ports <= 65535L),
+  all(ports >= 1024L),
+  length(ports) == length(truth),
+  all(ports == truth)
+)
+Sys.unsetenv("PORT4ME_TOOL")
+
+
+message('- port4me(user = "alice", tool = "jupyter-notebook")')
+truth <- 29525L
+ports <- port4me(user = "alice", tool="jupyter-notebook")
+print(ports)
+stopifnot(
+  is.integer(ports),
+  all(is.finite(ports)),
+  all(ports > 0L),
+  all(ports <= 65535L),
+  all(ports >= 1024L),
+  length(ports) == length(truth),
+  all(ports == truth)
+)
+
+message('- port4me() with PORT4ME_USER=alice')
+Sys.setenv(PORT4ME_USER = "alice")
+truth <- 30845L
+ports <- port4me()
 print(ports)
 stopifnot(
   is.integer(ports),
@@ -30,6 +104,100 @@ stopifnot(
   length(ports) == length(truth),
   all(ports == truth)
 )
+
+
+message('- port4me(user = "alice") with PORT4ME_LIST=10)')
+Sys.setenv(PORT4ME_LIST = "10")
+ports <- port4me()
+print(ports)
+stopifnot(
+  is.integer(ports),
+  all(is.finite(ports)),
+  all(ports > 0L),
+  all(ports <= 65535L),
+  all(ports >= 1024L),
+  length(ports) == length(truth),
+  all(ports == truth)
+)
+Sys.unsetenv("PORT4ME_LIST")
+
+
+exclude <- c(30845, 32310)
+message(sprintf("- port4me(exclude = c(%s))", paste(exclude, collapse = ", ")))
+port <- port4me(exclude = exclude)
+print(port)
+stopifnot(
+  length(port) == 1L,
+  is.integer(port),
+  is.finite(port),
+  port > 0L,
+  port <= 65535L,
+  port >= 1024L,
+  port == setdiff(truth, exclude)[1]
+)
+
+
+message(sprintf("- port4me() with PORT4ME_EXCLUDE=%s", paste(exclude, collapse = ", ")))
+Sys.setenv(PORT4ME_EXCLUDE = paste(exclude, collapse = ","))
+port <- port4me()
+print(port)
+stopifnot(
+  length(port) == 1L,
+  is.integer(port),
+  is.finite(port),
+  port > 0L,
+  port <= 65535L,
+  port >= 1024L,
+  port == setdiff(truth, exclude)[1]
+)
+Sys.unsetenv("PORT4ME_EXCLUDE")
+
+
+include <- c(2000:2123, 4321, 10000:10999)
+message("- port4me(include = c(2000:2123, 4321, 10000:10999))")
+port <- port4me(include = include)
+print(port)
+stopifnot(
+  length(port) == 1L,
+  is.integer(port),
+  is.finite(port),
+  port > 0L,
+  port <= 65535L,
+  port >= 1024L,
+  port == 10451L
+)
+
+
+include <- c(2000:2123, 4321, 10000:10999)
+message("- port4me() with PORT4ME_INCLUDE=...")
+Sys.setenv(PORT4ME_INCLUDE = paste(include, collapse = ","))
+port <- port4me()
+print(port)
+stopifnot(
+  length(port) == 1L,
+  is.integer(port),
+  is.finite(port),
+  port > 0L,
+  port <= 65535L,
+  port >= 1024L,
+  port == 10451L
+)
+Sys.unsetenv("PORT4ME_INCLUDE")
+
+
+prepend <- c(4321, 11001)
+message("- port4me(prepend = c(4321, 11001))")
+ports <- port4me(prepend = prepend, list = 5L)
+stopifnot(
+  is.integer(ports),
+  length(ports) == 5L,
+  all(is.finite(ports)),
+  all(ports > 0L),
+  all(ports <= 65535L),
+  all(ports >= 1024L),
+  all(ports == c(prepend, head(truth, n = 5L - length(prepend))))
+)
+
 
 n <- 200e3
 message(sprintf("- port4me(list = %d)", n))