+++ NEWS.md
@@ -1,3 +1,28 @@
+# Version 0.6.0 [2023-07-13]
+
+## New Features
+
+ * Add command-line interface for R, e.g. `Rscript -e
+   port4me::port4me`, `Rscript -e port4me::port4me --list=5`,
+   `Rscript -e port4me::port4me --test=8087`.
+
+## Miscellaneous
+
+ * Two of the examples for excluding ports had the wrong output.
+
+## Bug Fixes
+
+ * If argument `prepend` was a numeric, but not an integer, then the
+   return port, or ports, would also be numeric. Now `port4me()`
+   returns integer ports also when `prepend` is numeric.
+
+ * The R `port4me::port4me()` function did not default to environment
+   variable `PORT4ME_LIST`, if argument `list` was not specified.
+
+ * The R `port4me::port4me()` function did not default to environment
+   variable `PORT4ME_TEST`, if argument `test` was not specified.
+
+
 # Version 0.5.1 [2023-03-23]
 
 ## Miscellaneous